refactor: move pinned-communities next to its only consumer
The package sat at the repo root but is imported from exactly one place,
internal/protocol/pinnedcommunities/loader.go. It moves in beside it as
.../pinnedcommunities/assets, and the package is renamed to match its
directory so the import no longer needs an alias.

The embedded payload travels with it; the //go:embed pattern is relative
to the package directory, so it is unchanged.

refs #7067
