+++ internal/protocol/pinnedcommunities/assets/embedded.go
@@ -1,4 +1,4 @@
-package pinnedcommunitiesassets
+package assets
 
 import "embed"
